Renames zipkin projects to zipkin2 and removes "lc" stuff
"lc" was a hack for the v1 model. In v2 -> v1 conversion, we implicitly
create an empty "lc" when needed to attach a service name. By removing
the "local component" thing here, it is less cognitive load and more
like other integrations.
