+++ samples/sample-spring-boot/src/main/resources/application.properties
@@ -29,9 +29,3 @@ wingtips.zipkin.base-url=http://localhost:9411
 #   See the wingtips-zipkin readme (https://github.com/Nike-Inc/wingtips/tree/master/wingtips-zipkin) for details on
 #   how this service name is used.
 wingtips.zipkin.service-name=some-service-name
-
-# The Zipkin local component namespace for local-only spans, used when sending local-only Wingtips spans to Zipkin.
-#   If this property is not specified then "unknown" will be used.
-#   See the wingtips-kipkin readme (https://github.com/Nike-Inc/wingtips/tree/master/wingtips-zipkin) for details on
-#   how this local component namespace is used.
-wingtips.zipkin.local-component-namespace=some-local-component-name
