--- wingtips-zipkin-spring-boot/src/test/java/com/nike/wingtips/springboot/componenttest/componentscanonly/ComponentTestMainWithComponentScanOnly.java
@@ -1,13 +1,0 @@
-package com.nike.wingtips.springboot.componenttest.componentscanonly;
-
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
-
-@SpringBootApplication
-@ComponentScan(basePackages = "com.nike")
-public class ComponentTestMainWithComponentScanOnly {
-
-    public ComponentTestMainWithComponentScanOnly() {
-    }
-
-}
