+++ wingtips-zipkin2-spring-boot/src/test/java/com/nike/wingtips/springboot/zipkin2/componenttest/componentscanonly/ComponentTestMainWithComponentScanOnly.java
@@ -1,0 +1,13 @@
+package com.nike.wingtips.springboot.zipkin2.componenttest.componentscanonly;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
+
+@SpringBootApplication
+@ComponentScan(basePackages = "com.nike")
+public class ComponentTestMainWithComponentScanOnly {
+
+    public ComponentTestMainWithComponentScanOnly() {
+    }
+
+}
