+++ wingtips-zipkin2-spring-boot/src/test/java/com/nike/wingtips/springboot/zipkin2/componenttest/manualimportandcomponentscan/ComponentTestMainWithBothManualImportAndComponentScan.java
@@ -1,0 +1,17 @@
+package com.nike.wingtips.springboot.zipkin2.componenttest.manualimportandcomponentscan;
+
+import com.nike.wingtips.springboot.zipkin2.WingtipsWithZipkinSpringBootConfiguration;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Import;
+
+@SpringBootApplication
+@Import(WingtipsWithZipkinSpringBootConfiguration.class)
+@ComponentScan(basePackages = "com.nike")
+public class ComponentTestMainWithBothManualImportAndComponentScan {
+
+    public ComponentTestMainWithBothManualImportAndComponentScan() {
+    }
+
+}
